23042025 v3 FreeDayTypeService fixed that vacation (id=1) can not be erased, UserService fixed validation (can not click on validation two times), and OvertimeHoursService message for empty UserTotalAttendance changed

+++ src/main/java/com/caci/gaia_leave/administration/service/FreeDayTypeService.java
@@ -98,6 +98,9 @@ public class FreeDayTypeService {
         if (!freeDayTypeRepository.existsById(id)) {
             throw new CustomException("FreeDayType with id `" + id + "` doesn't exists.");
         }
+        if(id == 1){
+            throw new CustomException("FreeDayType with id `1` can not be deleted.");
+        }
 
         try{
             freeDayTypeRepository.deleteById(id);
